create RoomList component for better demo
